Roll our own client for func tests
Since we've already got a wrapper to represent questions, this
is pretty simple -- plus it lets us be a lot more precise in
our assertions, like seeing the CNAME records that are
returned in an A query.

At some point, we should maybe just drop dnpython as a test dep?